Compile time optimizations (#134)
* identify compile time optimizations

* apply compile time optimizations before cacheing execution plan

* address PR comments

* removed apply limit and apply skip optimizations

+++ src/execution_plan/ops/op_sort.h
@@ -16,9 +16,9 @@ typedef struct {
 	Record *buffer;        // Holds all records.
 	heap_t *heap;          // Holds top n records.
 	bool first;            // first visit to consume func
-	uint skip;             // Total number of records to skip
+	uint64_t skip;         // Total number of records to skip
 	uint record_idx;       // index of current record to return
-	uint limit;            // Total number of records to produce
+	uint64_t limit;        // Total number of records to produce
 	uint *record_offsets;  // All Record offsets containing values to sort by
 	int *directions;       // Array of sort directions(ascending / descending)
 	AR_ExpNode **exps;     // Projected expressons.
